task: remove public export of max precision util
