+++ packages/utils/src/formatNumber.ts
@@ -24,14 +24,3 @@ export function formatNumber(
   }
   return getFormatter(locale, options).format(value);
 }
-
-export function formatNumberMaxPrecision(
-  value: number | null,
-  locale?: Intl.LocalesArgument,
-  options?: Intl.NumberFormatOptions,
-) {
-  return formatNumber(value, locale, {
-    ...options,
-    maximumFractionDigits: 20,
-  });
-}
